Cocktail Sort Program in Python
Cocktail sort is the variation of Bubble Sort which traverses the list in both directions alternatively. It is different from a bubble sort in the sense that, bubble sort traverses the list in forwarding direction only, while this algorithm traverses in forward as well as the backward direction in one iteration.