Heap Sort Program in Python
Heap sort processes the elements by creating the min-heap or max-heap using the elements of the given array. Min heap or max heap represents the ordering of the array in which the root element represents the minimum or maximum element of the array. At each step, the root element of the heap gets deleted and stored into the sorted array and the heap will again be heaped.